pyproject.toml
[tool.poetry]
name = "crossword-generator"
version = "0.1.1"
description = "Swedish Crossword Generator with vector weighting (preference) of words"
authors = ["Adam Benda <adam@adasek.cz>"]
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
more-itertools = "=10.7.0"
pandas = "=2.3.0"
PyICU = "=2.15.2"
faktory = "=1.0.0"
python-dotenv = "=1.1.0"
requests = "=2.32.3"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
mypy = "^1.16.0"
isort = "^6.0.1"
pandas-stubs = "^2.2.3.250527"

[tool.pytest.ini_options]
pythonpath = [
    "."
]

[tool.mypy]
# Disallow dynamic typing
disallow_any_unimported = true
disallow_any_expr = true
disallow_any_decorated = true
disallow_any_generics = true
disallow_any_explicit = true
disallow_subclassing_any = true
files = "crossword"
exclude = [
    "experiment"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
